Add clippy and fix clippy errors
